requirements.txt
# HTTP client
requests==2.31.0

# GCP packages
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.1
google-cloud-firestore==2.14.0

# Environment variables
python-dotenv==1.0.1

# Date handling
pytz==2024.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Type checking
mypy==1.8.0

# Linting
black==24.1.1
flake8==7.0.0

# Logging
structlog==24.1.0